requirements.txt
gputil==1.3.0
colour==0.1.5
atomicwrites==1.1.5
attrs==18.1.0
avro-python3==1.8.2
boto3==1.9.35
botocore==1.12.31
certifi==2018.10.15
chardet==3.0.4
click==6.7
Cython==0.28.3
decorator==4.3.0
docopt==0.6.2
docutils==0.14
elasticsearch==6.3.0
fastavro==0.19.7
Flask==1.0.2
flask-cors
glog==0.3.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
more-itertools==4.2.0
networkx==2.1
numpy==1.14.5
pandas==0.23.1
Pillow==5.1.0
pipreqs==0.4.9
pluggy==0.6.0
py==1.5.3
pytest==3.6.2
python-dateutil==2.7.3
python-gflags==3.1.2
python-magic==0.4.15
pytz==2018.4
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-testadapter==0.3.0
rq==0.12.0
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
tabulate==0.8.2
urllib3==1.23
Werkzeug==0.14.1
yarg==0.1.9
matplotlib==2.2.2
colour==0.1.5
confluent-kafka==0.11.6
zipfile2==0.0.12
credstash==1.15.0
mysqlclient==1.3.13
tensorflow-gpu==1.10.1
tensorboard==1.10.0
